Implemented strength mode / removed "WIP"!
- a first version of the functionality i wanted to see.
- implemented strength mode, along with help and documentation
- made the makefile link the math library in the end
- fixed some int types
- reworded / fixed typos / fixed wording mistakes

+++ src/headers/interface/storage.h
@@ -13,7 +13,7 @@ void buildWordPoolFile(FILE* _source, FILE* _target, struct huffmanTree* _tree, 
 void restoreRawWordList(FILE* source, FILE* target, char* _alphabet, int _alphabetLength, bool verbose);
 
 // gets the amount of words stored in a binary
-unsigned int getWordPoolSize(FILE* _source);
+unsigned long int getWordPoolSize(FILE* _source);
 
 // reads the file and returns a list of the translated seeds. the seeds act as the indexes of words in the pool
 struct translatedSeedList* translateSeedListWithWordPool(
